+++ test/react-native/features/fixtures/app/react_native_navigation_js/install.sh
@@ -2,9 +2,7 @@
 
 npm i @bugsnag/plugin-react-native-navigation@$BUGSNAG_VERSION --registry=$REGISTRY_URL
 
-if [ "$REACT_NATIVE_VERSION" == "rn0.60" ]; then
-   npm i react-native-navigation@7.0.0
-elif [ "$REACT_NATIVE_VERSION" == "rn0.66" ]; then
+if [ "$REACT_NATIVE_VERSION" == "rn0.66" ]; then
    npm i react-native-navigation@7.29.1
 else
    npm i react-native-navigation@^7.30.0